Rebuilding run_int to not stitch two sims together, but rather for the model_strat to support parms_int and time_int
This is to fix an issue where we were seeing run_int and run_basic simulation outcomes differ because
run_int was stitching together two simulations.

What we saw happening was that simulations run with run_int that had exactly the same parameters for the
base case and for an intervention were differing from simulations run with run_basic run with the same parameters.

We believe this was happening because we think the ode function from deSolve does some interpolation.

Here's some literature describing a similar problem:
https://stackoverflow.com/questions/55961033/why-there-is-a-difference-in-ode-solution-using-event-function-in-r